Replace API with turbo-frame
