--- app/views/open_desk/menu/_top_menu_node.html.erb
@@ -1,1 +1,0 @@
-<%= angular_component_tag 'opce-opendesk-central-navigation' %>
