+++ config/routes.rb
@@ -1,0 +1,5 @@
+Rails.application.routes.draw do
+  namespace :open_desk do
+    get "/menu", to: "central_navigation#menu", as: :menu
+  end
+end
