--- frontend/module/central-navigation/central-navigation.model.ts
@@ -1,12 +1,0 @@
-export interface ILinkedApplication {
-  identifier:string;
-  name:string;
-  icon:string;
-  link:string;
-}
-
-export interface IApplicationGroup {
-  identifier:string;
-  name:string;
-  items:ILinkedApplication[];
-}
