+++ frontend/module/global_styles.scss
@@ -1,0 +1,1 @@
+@import "./styles/central_navigation";
