+++ frontend/module/styles/_central_navigation.sass
