+++ spec/lib/open_project/open_desk/central_navigation_adapters/open_desk_spec.rb
@@ -27,7 +27,7 @@
 #++
 require "spec_helper"
 
-RSpec.describe API::V3::LinkedApplications::Adapters::OpenDesk do
+RSpec.describe OpenProject::OpenDesk::CentralNavigationAdapters::OpenDesk do
   let(:login) { "foo" }
   let(:language) { "de" }
   let(:user) { build(:user, login:, language:) }
@@ -63,7 +63,7 @@ RSpec.describe API::V3::LinkedApplications::Adapters::OpenDesk do
   end
 
   describe "#fetch_entries" do
-    let(:fixture_path) { File.expand_path("../../../../../fixtures", __dir__) }
+    let(:fixture_path) { File.expand_path("../../../../fixtures", __dir__) }
     let(:logged_out_response) { File.read File.join(fixture_path, "logged_out_response.json") }
 
     subject { instance.fetch_entries }
